+++ src/fetch-options.ts
@@ -1,4 +1,4 @@
-import type { RetryConfiguring } from './fetch-retry/fetch-retry-configuring';
+import type { RetryConfiguring } from './fetch-retry/configuration/retry-configuring';
 
 export type FetchOptions = {
   requestInit?: RequestInit;
