+++ src/fetch-retry/configuration/default-retry-configuration.ts
@@ -1,14 +1,5 @@
-type TimeInterval = number; // milliseconds
-
-export interface RetryConfiguring {
-  shouldRetry(
-    response: Response | null,
-    retryNumber: number,
-    error?: unknown,
-  ): boolean;
-
-  retryDelay(retryNumber: number): TimeInterval;
-}
+import type { RetryConfiguring } from './retry-configuring';
+import type { TimeInterval } from './time-interval';
 
 export class DefaultRetryConfiguration implements RetryConfiguring {
   private readonly maxRetries: Readonly<number> = 2;
@@ -28,15 +19,5 @@ export class DefaultRetryConfiguration implements RetryConfiguring {
   retryDelay(retryNumber: number): TimeInterval {
     // Exponential backoff up to 10 seconds
     return Math.min(500 * 2 ** retryNumber, 10000);
-  }
-}
-
-export class NoRetryConfiguration implements RetryConfiguring {
-  shouldRetry(): boolean {
-    return false;
-  }
-
-  retryDelay(): TimeInterval {
-    return 0;
   }
 }
