+++ src/fetch-retry/configuration/no-retry-configuration.ts
@@ -1,0 +1,12 @@
+import type { RetryConfiguring } from './retry-configuring';
+import type { TimeInterval } from './time-interval';
+
+export class NoRetryConfiguration implements RetryConfiguring {
+  shouldRetry(): boolean {
+    return false;
+  }
+
+  retryDelay(): TimeInterval {
+    return 0;
+  }
+}
