+++ src/fetch-retry/configuration/time-interval.ts
@@ -1,0 +1,1 @@
+export type TimeInterval = number; // milliseconds
